Rename package autotester to tester
closes #101
