+++ tester/src/main/scala/com/easyforger/tester/uTestRunnerMod.scala
@@ -2,7 +2,7 @@
  * This file is part of EasyForger which is released under GPLv3 License.
  * See file LICENSE.txt or go to http://www.gnu.org/licenses/gpl-3.0.en.html for full license details.
  */
-package com.easyforger.autotester
+package com.easyforger.tester
 
 import com.easyforger.base.{EasyForger, LootTableLoadEventReplacerTest}
 import com.easyforger.creatures.CreaturesHandlerTest
